+++ src/native/cipher.rs
@@ -1,5 +1,3 @@
-use std::{convert::TryInto, str::FromStr};
-
 use crate::{
     fawkes_crypto::{
         ff_uint::{Num, seedbox::{SeedboxChaCha20, SeedBox, SeedBoxGen}},
@@ -22,27 +20,7 @@ use chacha20poly1305::{ChaCha20Poly1305, Key, Nonce, aead::AeadMutInPlace};
 use chacha20poly1305::aead::{Aead, NewAead};
 use chacha20poly1305::aead::heapless::Vec as HeaplessVec;
 
-
-#[derive(Serialize, Deserialize, PartialEq, Debug, Clone, Copy, Default)]
-pub enum ECPointsFormat {
-    #[default]
-    XCoordinate,
-    XCoordinateWithSign,
-    XYCoordinates,
-}
-
-impl FromStr for ECPointsFormat {
-    type Err = ();
-
-    fn from_str(input: &str) -> Result<ECPointsFormat, Self::Err> {
-        match input {
-            "XCoordinate" => Ok(ECPointsFormat::XCoordinate),
-            "XCoordinateWithSign" => Ok(ECPointsFormat::XCoordinateWithSign),
-            "XYCoordinates" => Ok(ECPointsFormat::XYCoordinates),
-            _ => Err(()),
-        }
-    }
-}
+use super::ec_points::{ECPointsFormat, skip_ec_point, parse_ec_point, check_in_prime_subgroup};
 
 /// Wrapper for HeaplessVec (if buffer size is less or equals to N) or Vec otherwise
 enum Buffer<T, const N: usize> {
@@ -278,56 +256,12 @@ pub fn decrypt_in<P: PoolParams>(eta: Num<P::Fr>, memo: &[u8], params: &P, ec_po
     }
 }
 
-fn buf_take<'a>(memo: &mut &'a[u8], size:usize) -> Option<&'a[u8]> {
+pub(crate) fn buf_take<'a>(memo: &mut &'a[u8], size:usize) -> Option<&'a[u8]> {
     if memo.len() < size {
         None
     } else {
         let res = &memo[0..size];
         *memo = &memo[size..];
         Some(res)
-    }
-}
-
-fn parse_ec_point<'a, P: PoolParams>(memo: &mut &'a[u8], num_size: usize, params: &P, ec_points_format: ECPointsFormat) -> Option<EdwardsPoint<P::Fr>> {
-    match ec_points_format {
-        ECPointsFormat::XCoordinate => {
-            EdwardsPoint::subgroup_decompress(Num::deserialize(memo).ok()?, params.jubjub())
-        },
-        ECPointsFormat::XCoordinateWithSign => {
-            let bytes = buf_take(memo, num_size)?;
-            EdwardsPoint::decompress_unchecked(bytes.try_into().ok()?, params.jubjub())
-        },
-        ECPointsFormat::XYCoordinates => {
-            let point = EdwardsPoint{ x: Num::deserialize(memo).ok()?, y: Num::deserialize(memo).ok()? };
-            if point.is_in_curve(params.jubjub()) {
-                Some(point)
-            } else {
-                None
-            }            
-        }
-    }
-}
-
-fn skip_ec_point<'a>(memo: &mut &'a[u8], num_size: usize, ec_points_format: ECPointsFormat) -> Option<()> {
-    match ec_points_format {
-        ECPointsFormat::XYCoordinates => {
-            buf_take(memo, num_size * 2)?;
-        },
-        _ => {
-            buf_take(memo, num_size)?;
-        }
-    } 
-    Some(())    
-}
-
-fn check_in_prime_subgroup<P: PoolParams>(p: EdwardsPoint<P::Fr>, params: &P, ec_points_format: ECPointsFormat) -> Option<()> {
-    match ec_points_format {
-        ECPointsFormat::XCoordinate => {
-            // we've already checked it in subgroup_decompress
-            Some(())
-        },
-        _ => {
-           p.is_in_prime_subgroup(params.jubjub()).then_some(())
-        }
     }
 }
