+++ src/native/mod.rs
@@ -7,6 +7,7 @@ pub mod cipher;
 pub mod key;
 pub mod tree;
 pub mod delegated_deposit;
+pub mod ec_points;
 
 mod borsh;
 mod sample;
